UID is now integer

+++ db/migrate/20111201182811_change_authentications_uid_to_integer.rb
@@ -1,5 +1,5 @@
 class ChangeAuthenticationsUidToInteger < ActiveRecord::Migration
   change_table :authentications do |t|
-    t.change :uid, :bigint
+    t.change :uid, :integer
   end
 end
